\docType{data}
\name{Mmus_example_lis}
\title{M3D Example Data}
\alias{Mmus_example_list}
\format{A list with two elements: \code{data} and \code{labels}}
\usage{Deng_subset_example_data}
\description{A subset of the Deng et al. (2014) single-cell RNA-Seq expression data for mouse 2-cell embryos and blastocysts. Gene names are HGNC symbols. labels are sampling timepoints.}
\value{A list with two elements: \code{data} and \code{labels}}
\keyword{datasets}

